Unify recommendation language and syntax
Sometimes we state a recommendations with "Always...." "Don't..." ... etc.

We should standardize the way we do recommendations. IMHO, the best way would be to do imperative statements starting with a verb. Additionally, "always" is redundant, as we should *always* follow the recommendations or make it clear that something should be done "whenever possible". Similarly, instead of "Don't..." we can instead start with the imperative "Avoid...". 

Finally, all references to code should be highlighted.